+++ dccd/continuous_dl/__init__.py
@@ -28,10 +28,12 @@ exchanges.*
 # Third party packages
 
 # Local packages
-from . import bitfinex, bitmex, exchange
+from . import bitfinex, bitmex, bybit, exchange
 from .bitfinex import *
 from .bitmex import *
+from .bybit import *
 
 __all__ = ['exchange']
 __all__ += bitfinex.__all__
 __all__ += bitmex.__all__
+__all__ += bybit.__all__